requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch==2.3.0
torchvision==0.18.0
pydantic==2.5.3
python-multipart==0.0.6
Pillow==10.2.0
numpy==1.26.4
